DQA-4743: Move run to root & adapt tests.
