+++ composer.json
@@ -10,6 +10,7 @@
         "email": "DIGIT-NEXTEUROPA-QA@ec.europa.eu",
         "source": "https://github.com/ec-europa/toolkit"
     },
+    "bin": ["run"],
     "require": {
         "php": ">=8.1",
         "ext-curl": "*",
@@ -17,7 +18,7 @@
         "ext-json": "*",
         "ext-simplexml": "*",
         "behat/behat": "^3.7 || ^3.8",
-        "consolidation/robo": "^4.0",
+        "consolidation/robo": "^3.0 || ^4.0",
         "cweagans/composer-patches": "^1.4 || ^1.7",
         "drush/drush": "^9.7.1 || ^10.0.0 || ^11.0.4",
         "ec-europa/qa-automation": "^9.0",
@@ -60,6 +61,5 @@
         "post-update-cmd": [
             "./vendor/bin/phpcs --config-set installed_paths '../../drupal/coder/coder_sniffer,../../phpcompatibility/php-compatibility,phpcs,../../ec-europa/qa-automation/phpcs'"
         ]
-    },
-    "bin": ["bin/run"]
+    }
 }
