+++ tests/Commands/InstallCommandsTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace EcEuropa\Toolkit\Tests\Commands;
 
+use EcEuropa\Toolkit\TaskRunner\Runner;
 use EcEuropa\Toolkit\Tests\AbstractTest;
-use OpenEuropa\TaskRunner\TaskRunner;
 use Symfony\Component\Console\Input\StringInput;
 use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Yaml\Yaml;
@@ -49,7 +49,7 @@ class InstallCommandsTest extends AbstractTest
         // Run command.
         $input = new StringInput($command . ' --simulate --working-dir=' . $this->getSandboxRoot());
         $output = new BufferedOutput();
-        $runner = new TaskRunner($input, $output, $this->getClassLoader());
+        $runner = new Runner($this->getClassLoader(), $input, $output);
         $runner->run();
 
         // Assert expectations.
